IGNITE-23114 Code cleanup, javadocs

+++ modules/core/src/main/java/org/apache/ignite/internal/jdbc/thin/ConnectionProperties.java
@@ -579,5 +579,5 @@ public interface ConnectionProperties {
      *
      * @param maxInMemoryLobSize Maximum size in bytes.
      */
-    public void setMaxInMemoryLobSize(int maxInMemoryLobSize) throws SQLException;
+    public void setMaxInMemoryLobSize(long maxInMemoryLobSize) throws SQLException;
 }
